Add Ceph + Rex Vagrant environment
This introduces a Vagrant environment for demo'ing REX-Ray with Ceph
RBD. It can handle deploying a Ceph environment of 1 or more nodes. A
custom vagrant box is built via Packer in order to minimize package
downloads. REX-Ray is pre-configured on the ceph-admin node. RBD
mounts should never be done via kernel module on the same nodes running
OSDs.

REX-Ray is pre-installed, but can be updated by uncommenting lines in the
Vagrant file. The pre-installed rexray right now doesn't have the RBD
driver in it, so the Vagrantfile currently defaults to building rexray
from source via the appropriate developer branch.