Use the vagrant metadata file for box_url
Point the JSON metadata file as the box_url. We use rawgit.com so
that the Content Header comes back as application/json. This makes
GitHub work exactly like Atlas in this context. The box_url should
not need to be updated again, but we can add box versions to the
metadata file as we release them.