Vagrant 1.9.1 private networking workaround
Vagrant 1.9.1 has an issue where private networks on Red Hat flavored
distros are not brought up correctly. The interface is "ifdown"'d, and
then NetworkManager is restarted, but the private network interface is
specifcally configured to not be controlled by NetworkManger --
therefore the interface is never brought back up.

Workaround that by issuing a "service network restart" before we start
doing things that require that network interface.