updated guide and script and Vagrantfile
updated the Vagrant file to no longer build from source since it's now part of the stable release. Changed the script to run rexray as a service instead of invoking systemctl since it hasn't been installed as a service. updated the guide to better reflect the labs way of doing tutorials.

Signed-off-by: Kendrick Coleman <kendrickcoleman@gmail.com>