added mesos install option to ScaleIO
the new mesos install option for ScaleIO allows us to easily demonstrate the usage of REX-Ray and Docker with Mesos/Marathon. Everything is completely setup to start curling applications at the Marathon endpoint of 192.168.50.12. all other mesos installs have been moved to the old folder for now. This is done because those projects are no longer being maintained and are very out of date. By default, the latest versions of docker/rexray/mesos/marathon are all installed.

Signed-off-by: kacole2 <kendrickcoleman@gmail.com>